+++ src/biz/bokhorst/xprivacy/ActivityMain.java
@@ -1,7 +1,6 @@
 package biz.bokhorst.xprivacy;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;