+++ src/biz/bokhorst/xprivacy/PrivacyManager.java
@@ -10,7 +10,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -487,7 +486,9 @@ public class PrivacyManager {
 		if (rCursor != null)
 			try {
 				final int max = rCursor.getCount();
-				final int step = (max + 95)/96; // 96% left for loading the restrictions
+				final int step = (max + 95) / 96;
+				// 96% left for loading the restrictions
+
 				int current = 0;
 				while (rCursor.moveToNext()) {
 					current++;
@@ -506,65 +507,6 @@ public class PrivacyManager {
 				rCursor.close();
 			}
 		return result;
-	}
-
-	public static class RestrictedIterator implements Iterator<RestrictionDesc>, Iterable<RestrictionDesc> {
-
-		private Cursor mCursor;
-
-		public RestrictedIterator(Cursor cursor) {
-			mCursor = cursor;
-		}
-
-		public int getCount() {
-			return mCursor != null ? mCursor.getCount() : 0;
-		}
-
-		public void cleanUp() {
-			try {
-				mCursor.close();
-			} finally {
-				// nothing to do here
-			}
-		}
-
-		@Override
-		public boolean hasNext() {
-			if (mCursor.moveToNext()) {
-				return true;
-			} else {
-				cleanUp();
-				return false;
-			}
-		}
-
-		@Override
-		public RestrictionDesc next() {
-			RestrictionDesc restriction = new RestrictionDesc();
-			restriction.uid = mCursor.getInt(mCursor.getColumnIndex(PrivacyProvider.COL_UID));
-			restriction.restricted = Boolean.parseBoolean(mCursor.getString(mCursor
-					.getColumnIndex(PrivacyProvider.COL_RESTRICTED)));
-			restriction.restrictionName = mCursor.getString(mCursor
-					.getColumnIndex(PrivacyProvider.COL_RESTRICTION));
-			restriction.methodName = mCursor.getString(mCursor.getColumnIndex(PrivacyProvider.COL_METHOD));
-			return restriction;
-		}
-
-		@Override
-		public void remove() {
-			throw new UnsupportedOperationException("remove not supported");
-		}
-
-		@Override
-		public Iterator<RestrictionDesc> iterator() {
-			return this;
-		}
-	}
-
-	public static RestrictedIterator getRestrictedIterator(Context context) {
-		Cursor rCursor = context.getContentResolver().query(PrivacyProvider.URI_RESTRICTION, null, null,
-				new String[] { Integer.toString(0), Boolean.toString(false) }, null);
-		return new RestrictedIterator(rCursor);
 	}
 
 	public static void deleteRestrictions(Context context, int uid) {
@@ -747,7 +689,7 @@ public class PrivacyManager {
 				int current = 0;
 				while (sCursor.moveToNext()) {
 					current++;
-					if (current == max/2 || current == max)
+					if (current == max / 2 || current == max)
 						progress.run(); // 2% for fetching settings
 					// Get setting
 					String setting = sCursor.getString(sCursor.getColumnIndex(PrivacyProvider.COL_SETTING));
@@ -758,65 +700,6 @@ public class PrivacyManager {
 				sCursor.close();
 			}
 		return result;
-	}
-
-	public static class SettingDesc {
-		public String settingName;
-		public String value;
-	}
-
-	public static class SettingsIterator implements Iterator<SettingDesc>, Iterable<SettingDesc> {
-
-		private Cursor mCursor;
-
-		public SettingsIterator(Cursor cursor) {
-			mCursor = cursor;
-		}
-
-		public int getCount() {
-			return mCursor != null ? mCursor.getCount() : 0;
-		}
-
-		public void cleanUp() {
-			try {
-				mCursor.close();
-			} finally {
-				// nothing to do here
-			}
-		}
-
-		@Override
-		public boolean hasNext() {
-			if (mCursor.moveToNext()) {
-				return true;
-			} else {
-				cleanUp();
-				return false;
-			}
-		}
-
-		@Override
-		public SettingDesc next() {
-			SettingDesc setting = new SettingDesc();
-			setting.settingName = mCursor.getString(mCursor.getColumnIndex(PrivacyProvider.COL_SETTING));
-			setting.value = mCursor.getString(mCursor.getColumnIndex(PrivacyProvider.COL_VALUE));
-			return setting;
-		}
-
-		@Override
-		public void remove() {
-			throw new UnsupportedOperationException("remove not supported");
-		}
-
-		@Override
-		public Iterator<SettingDesc> iterator() {
-			return this;
-		}
-	}
-
-	public static SettingsIterator getSettingsIterator(Context context) {
-		Cursor sCursor = context.getContentResolver().query(PrivacyProvider.URI_SETTING, null, null, null, null);
-		return new SettingsIterator(sCursor);
 	}
 
 	public static void deleteSettings(Context context) {